app_examples/novel/backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
httpx>=0.25.0
pyyaml>=6.0
simpleaudio>=1.0.4
pillow>=10.0.0
mutagen>=1.47.0
python-multipart>=0.0.6